Ship Alone Attribution to RefField5 if sku falls in eligable dept.
Items on the inbound CO that are in the departments listed in the yaml
property are flagged as SA or ship alone.  Also, added debugging you can
comment out.  On feature of the logic is that if you have a SA order
line, it won't check the DCX and VDCX eligablity.  After talking with
Jason, and Trevor, we should make them mutually exclusive.